# src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Eventos",
    "version": "1.0.0",
    "description": "Documentação interativa da API de Gerenciamento de Eventos (Eventos, Cronogramas, Convidados, Presença e Fotos)."
  },
  "servers": [
    {
      "url": "{baseUrl}/api",
      "description": "Servidor Atual",
      "variables": {
        "baseUrl": {
          "default": "http://localhost:8752"
        }
      }
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "ApiResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" },
          "message": { "type": "string" }
        }
      },
...
      "PaginatedResponse": {
        "type": "object",
        "properties": {
          "items": { "type": "array", "items": { "type": "object" } },
          "totalItems": { "type": "integer" },
          "totalPages": { "type": "integer" },
          "currentPage": { "type": "integer" },
          "limit": { "type": "integer" }
        }
      }
    }
  },
  "paths": {
...
    "/events/{eventId}/confirmations/all": {
      "get": {
        "summary": "Listar Relatório de Todas as Confirmações do Evento",
        "tags": ["Confirmations"],
        "parameters": [
          { "in": "path", "name": "eventId", "required": true, "schema": { "type": "integer" } },
          { "in": "query", "name": "page", "schema": { "type": "integer", "default": 1 } },
          { "in": "query", "name": "limit", "schema": { "type": "integer", "default": 10 } }
        ],
        "responses": {
          "200": { 
            "description": "Lista unificada paginada",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ApiResponse" },
                    { 
                      "type": "object", 
                      "properties": { 
                        "data": { 
                          "allOf": [
                            { "$ref": "#/components/schemas/PaginatedResponse" },
                            { "type": "object", "properties": { "items": { "type": "array", "items": { "type": "object" } } } }
                          ]
                        } 
                      } 
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
...
    "/events/{eventId}/gallery": {
      "get": {
        "summary": "Listar Fotos da Galeria do Evento",
        "tags": ["Gallery"],
        "parameters": [
          { "in": "path", "name": "eventId", "required": true, "schema": { "type": "integer" } },
          { "in": "query", "name": "page", "schema": { "type": "integer", "default": 1 } },
          { "in": "query", "name": "limit", "schema": { "type": "integer", "default": 10 } }
        ],
        "responses": {
          "200": { 
            "description": "Galeria paginada",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ApiResponse" },
                    { 
                      "type": "object", 
                      "properties": { 
                        "data": { 
                          "allOf": [
                            { "$ref": "#/components/schemas/PaginatedResponse" },
                            { "type": "object", "properties": { "items": { "type": "array", "items": { "type": "object" } } } }
                          ]
                        } 
                      } 
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  }
}